Added Epoch dayz_code
Added Epoch DayZ_Code to the repository as some of the files are needed
for the mods.  Not for running on the SERVER this is just here as a
placeholder for you to copy files over to mods that need them, for
example:

If you need to make custom files usch as: compiles.sqf, varibles.sqf and
fn_selfActions.sqf you will find these files in the mission PBO as there
are a lot of customizations that are made to these files to get mods for
Epoch to work.